+++ dependency-reduced-pom.xml
@@ -3,7 +3,7 @@
   <modelVersion>4.0.0</modelVersion>
   <groupId>com.warterpl</groupId>
   <artifactId>minecartannouncer</artifactId>
-  <version>2.0.0</version>
+  <version>2.0.1</version>
   <build>
     <plugins>
       <plugin>
