+++ pom.xml
@@ -5,7 +5,7 @@
 
     <groupId>com.warterpl</groupId>
     <artifactId>minecartannouncer</artifactId>
-    <version>2.0.0</version>
+    <version>2.0.1</version>
     <packaging>jar</packaging>
