+++ src/main/java/com/warterpl/helper/Utils.java
@@ -1,0 +1,9 @@
+package com.warterpl.helper;
+
+import org.bukkit.block.Block;
+
+public class Utils {
+    public static String getBlockKey(Block block) {
+        return block.getWorld().getName() + ";" + block.getX() + ";" + block.getY() + ";" + block.getZ();
+    }
+}
