+++ src/main/java/com/warterpl/minecartannoucer/Commands/DevCmdexec.java
@@ -17,14 +17,14 @@ public class DevCmdexec
             if (MinecartAnnouncer.msgBlocks.isEmpty()) {
                 player.sendMessage("There aren't any assigned messages.");
             } else {
-                player.sendMessage("Assigned messages:");
+                player.sendMessage("Assigned messages are on:");
                 for (var entry : MinecartAnnouncer.msgBlocks) {
                     player.sendMessage("> " + entry.getLocation());
                 }
             }
             return true;
         }
-        commandSender.sendMessage("Only players can execute that command.");
+        commandSender.sendMessage("Only players with op privileges can execute that command.");
         return false;
     }
 }
